requirements.txt
# APE 2026 - Python Dependencies
# Week 1: Foundation Setup

# ==============================================================================
# Core Framework
# ==============================================================================
python-dotenv==1.0.0          # Environment variables
pydantic==2.6.0               # Data validation
pydantic-settings==2.1.0      # Settings management

# ==============================================================================
# Vector Store (Week 1 Day 2)
# ==============================================================================
chromadb==0.4.22              # Embedded vector database
sentence-transformers==2.3.1  # Embedding generation
# Note: ChromaDB will auto-install: onnxruntime, tokenizers, huggingface-hub

# ==============================================================================
# LLM Providers (Week 1 Day 3)
# ==============================================================================
anthropic==0.18.1             # Claude API (Sonnet 4.5 for PLAN node)
openai==1.12.0                # GPT-4.5 for Doubter (Week 8) + DeepSeek API
google-generativeai==0.8.3   # Gemini 2.0 for LLM Debate (Week 11)

# ==============================================================================
# Database Drivers
# ==============================================================================
# Neo4j (Week 5-6)
neo4j==5.17.0                 # Graph database driver

# TimescaleDB / PostgreSQL (Week 1 Day 1 - already running)
psycopg2-binary==2.9.9        # PostgreSQL adapter
sqlalchemy==2.0.27            # ORM for TimescaleDB
alembic==1.13.1               # Database migrations

# Redis (Week 1 Day 1 - already running)
redis==5.0.1                  # Redis client

# ==============================================================================
# Orchestration (Week 9-10)
# ==============================================================================
langgraph==0.0.40             # LangGraph state machine
langchain==0.1.9              # LangChain core
langchain-anthropic==0.1.4    # Anthropic integration
langchain-openai==0.0.6       # OpenAI integration

# ==============================================================================
# Data Sources (Week 2+)
# ==============================================================================
yfinance==0.2.36              # Yahoo Finance API
fredapi==0.5.1                # FRED (Federal Reserve) API
pandas==2.2.0                 # Data manipulation
numpy==1.26.4                 # Numerical computing

# ==============================================================================
# API Framework (Week 13)
# ==============================================================================
fastapi==0.109.2              # REST API framework
uvicorn[standard]==0.27.1     # ASGI server
python-multipart==0.0.9       # Form data handling
python-jose[cryptography]==3.3.0  # JWT tokens
passlib[bcrypt]==1.7.4        # Password hashing
slowapi==0.1.9                # Rate limiting

# ==============================================================================
# Testing (Week 1+)
# ==============================================================================
pytest==8.0.0                 # Test framework
pytest-asyncio==0.23.5        # Async test support
pytest-cov==4.1.0             # Coverage reporting
pytest-mock==3.12.0           # Mocking support
httpx==0.26.0                 # Async HTTP client for testing

# ==============================================================================
# Development Tools
# ==============================================================================
ruff==0.2.2                   # Fast linter
black==24.2.0                 # Code formatter
mypy==1.8.0                   # Type checking
pre-commit==3.6.2             # Git hooks

# ==============================================================================
# Monitoring (Week 14)
# ==============================================================================
prometheus-client==0.20.0     # Metrics exporter
structlog==24.1.0             # Structured logging

# ==============================================================================
# Security (Week 15)
# ==============================================================================
bandit==1.7.7                 # Security linter
safety==3.0.1                 # Dependency vulnerability scanner

# ==============================================================================
# VEE Sandbox (Week 2)
# ==============================================================================
docker==7.0.0                 # Docker SDK for Python
watchdog==4.0.0               # Filesystem monitoring

# ==============================================================================
# Version Constraints
# ==============================================================================
# Python >= 3.11 (for performance + type hints)
# ChromaDB requires Python 3.8-3.11 (not 3.12 yet as of 0.4.22)
